requirements.txt
pandas
plotly
streamlit
numpy
pydtmc
pyyaml
ipdb
